(BSR)[API] chore: Move partial indexation commands to `core/search` package
